Add Tranquility to the mix
Unfortunately I can't figure out how to configure the ZooKeeper address on the command line, it seems like it has to be specified in the configuration file. This means that you won't be able to run the stack without creating a Tranquility log file.

The S3 extension must also be removed from Druid, otherwise the indexer tasks will crash with "Unable to load AWS credentials from any provider in the chain". Good work.