Attempt another approach to configuration: files
The strategy I've tried so far has been to use the default config that ships with Druid and disable the properties I want to change, so that I can set them as command line flags. This became pretty messy after a while, so this is an attempt to do the reverse: have the config outside of the containers and mount it.

This will make it easier to tweak both JVM parameters, logging config and Druid configuration.

I've also removed the Middle Manager and configured the Overlord to do that job.